+++ app/src/scenes/contact/components/PublicContactForm.jsx
@@ -56,8 +56,7 @@ export default function PublicContactForm() {
           const translationKey = filesResponse.code === "FILE_SCAN_DOWN" ? "FILE_SCAN_DOWN_SUPPORT" : filesResponse.code;
           return toastr.error("Une erreur s'est produite lors de l'upload des fichiers :", translate(translationKey), { timeOut: 5000 });
         }
-        uploadedFiles = filesResponse.uploadFiles;
-        console.log("🚀 ~ file: PublicContactForm.jsx:59 ~ handleSubmit ~ uploadedFiles:", uploadedFiles);
+        uploadedFiles = filesResponse.data;
       }
 
       const response = await API.post("/zammood/ticket/form", {